refactor: Update import paths for RootProvider and AuthProvider
