+++ frontend/src/app/(auth)/register/page.tsx
