+++ frontend/src/app/(main)/layout.tsx
@@ -1,5 +1,6 @@
 import type { Metadata, Viewport } from 'next';
 import { Inter } from 'next/font/google';
+import MainLayout from './MainLayout';
 
 const inter = Inter({ subsets: ['latin'] });
 
@@ -20,9 +21,5 @@ export default function Layout({
 }: Readonly<{
   children: React.ReactNode;
 }>) {
-  return (
-    <html lang="en" suppressHydrationWarning>
-      <body className={inter.className}>{children}</body>
-    </html>
-  );
+  return <MainLayout>{children}</MainLayout>;
 }
